backend/requirements.txt
fastapi
uvicorn[standard]
psutil
pydantic
pyyaml
